- Added image generation - Separated Export function to allow grabbing data independently - Adjusted "Builds" to "Loadout" on button text - Added sleep deprivation (being patched rn)
